fix: update result conversion from Long to Int in Question5.kt

+++ app/src/main/java/com/pohnpawit/persec2026/question5/Question5.kt
@@ -14,7 +14,7 @@ fun solveQuestion5(input: Int): Int {
         .toList()
         .sortedDescending()
         .joinToString("")
-        .toLong()
+        .toInt()
 
     require(result <= Int.MAX_VALUE) { "result exceeds int range: $result" }
     return result.toInt()
